pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vantage6-strongaya-rdf"
version = "1.0.0"
description = "A small library with various Resource Description Framework (RDF) data tools for Vantage6 algorithms developed for the STRONG AYA Project."
authors = [
    { name = "Joshi Hogenboom (@Jhogenboom)", email = "joshi.hogenboom@maastrichtuniversity.nl" },
    { name = "Varsha Gouthamchand (@VarshaGouthamchand)", email = "varsha.gouthamchand@maastro.nl" }
]
maintainers = [
    { name = "Joshi Hogenboom (@Jhogenboom)", email = "joshi.hogenboom@maastrichtuniversity.nl" },
    { name = "Varsha Gouthamchand (@VarshaGouthamchand)", email = "varsha.gouthamchand@maastro.nl" }
]

readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "pandas>=2.2.3",
    "vantage6-algorithm-tools>=3.6.0",
    "requests",
    "vantage6-strongaya-general @ git+https://github.com/STRONGAYA/v6-tools-general.git@v1.0.2"
]

[project.urls]
"Homepage" = "https://github.com/STRONGAYA/v6-tools-rdf"
"Bug Tracker" = "https://github.com/STRONGAYA/v6-tools-rdf/issues"

[tool.setuptools]
package-dir = { "" = "src" }
packages = ["vantage6_strongaya_rdf"]
package-data = { "vantage6_strongaya_rdf" = ["query_templates/*.rq"] }

